+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.0] - 2021-08-29
+### Changed
+- Now using youtube-scrapper and ytdl-core-discord instead of yt-search and ytdl-core.
+
+### Fixed
+- Error message being undefined when requesting an invalid SoundCloud track link.
+
 ## [0.0.0] - 2021-08-29
 ### Added
 - Author to `package.json`.
@@ -8,10 +15,11 @@ All notable changes to this project will be documented in this file.
 
 ## [0.0.0-13.1.0] - 2021-08-24
 ### Changed
-- Now use discord.js v13.1.0 instead of discord.js v12.5.3.
+- Now using discord.js v13.1.0 instead of discord.js v12.5.3.
 
 ## [0.0.0-12.5.3] - 2021-08-24
 
+[0.1.0]: https://github.com/joaompfonseca/panda-bot/compare/v0.0.0...v0.1.0
 [0.0.0]: https://github.com/joaompfonseca/panda-bot/compare/v0.0.0-13.1.0...v0.0.0
 [0.0.0-13.1.0]: https://github.com/joaompfonseca/panda-bot/compare/v0.0.0-12.5.3...v0.0.0-13.1.0
 [0.0.0-12.5.3]: https://github.com/joaompfonseca/panda-bot/releases/tag/v0.0.0-12.5.3